requirements.txt
numpy>=2.0.0
python-dotenv>=1.0.0
openai>=2.0.0
scikit-learn>=1.0.0

